Move helper methods to own file
