+++ config/application.rb
@@ -1,30 +1,11 @@
 require_relative "boot"
+require_relative "subscriptions"
 
 require "rails/all"
 
 # Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
 Bundler.require(*Rails.groups)
-
-def stripe_price_id(subscription)
-  credentials = Rails.application.credentials
-  if Rails.env.development?
-    credentials.dig(:stripe, :price_ids, subscription) ||
-      "#{subscription}_dummy_stripe_price_id"
-  else
-    credentials.stripe[:price_ids][subscription]
-  end
-end
-
-def revenue_cat_product_identifier(subscription)
-  credentials = Rails.application.credentials
-  if Rails.env.development?
-    credentials.dig(:revenue_cat, :product_identifiers, subscription) ||
-      "#{subscription}_dummy_revenue_cat_product_identifier"
-  else
-    credentials.revenue_cat[:product_identifiers][subscription]
-  end
-end
 
 module Railsdevs
   class Application < Rails::Application
